Refactor GitHub Actions Workflow Environment Variables
Moved the environment variable block to the top of the job
definition in the GitHub Actions workflow. This change improves
the organization of the workflow file, making it clearer which
environment variables are used throughout the pipeline.
Additionally, it reduces redundancy by ensuring that all steps
can access the necessary secrets without needing to redefine
them in each step.